+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.0.3] (2023-10-16)
+
+#### Improvements
+
+-   Added support for passing an array of styles.
+-   Removed misc that was earlier used for debug.
+-   Added image placeholder when it can not be loaded from uri.
+
+#### Changes
+
+-   Changed default of `resizeMode` from contain to cover.
+
 ## [0.0.2] (2023-10-16)
 
 #### Added
@@ -11,3 +23,4 @@
 -   Initial release.
 
 [0.0.2]: https://github.com/xerdnu/react-native-blasted-image/compare/v0.0.1...v0.0.2
+[0.0.3]: https://github.com/xerdnu/react-native-blasted-image/compare/v0.0.2...v0.0.3